// src/store/orders/orders.enum.ts
enum OrdersActionTypeEnum {
  ORDERS_LOADING = "ORDERS_LOADING",
  ORDERS_LOADED = "ORDERS_LOADED",
  
  GET_ORDER_ID = "GET_ORDER_ID",
  GET_ORDER_BY_ID = "GET_ORDER_BY_ID",

  GET_ORDERS_FUEL = "GET_ORDERS_FUEL",
  GET_ORDERS_GAS = "GET_ORDERS_GAS",

  CANCEL_ORDER = "CANCEL_ORDER",
  
  RESCHEDULE_TIME_SLOTS = "RESCHEDULE_TIME_SLOTS",
  RESCHEDULE_ORDER_UPDATE = "RESCHEDULE_ORDER_UPDATE",
}

export default OrdersActionTypeEnum;